Fixing the attach disk call
When attaching a disk, the CPI was not preserving all the agent
configuration, including the `update_settings.json` file, this was
causing problems when attaching disks to VMs that contained nats
credentials in the update_settings file.

This commit will change the attach disk process, in addition to creating
a new VM and copying `warden-cpi-agent-env.json` it will create the VM
based on a snapshot of the current VM.

Attach disk was failing due to update_settings.json not being copied.

Signed-off-by: Daniel Felipe Ochoa <danielfelipo@vmware.com>