Revert "Fixing the attach disk call"
This reverts commit 230365256050f2d59f0b28455064c6d208ac654e.

Taking a snapshot of the VM seemed to cause problems. For some reason the bosh
director fails to start. Monit is unable to call the http health check for the
Health Monitor job, possibly related to iptable rules that are installed in the
first version of the "VM", but then don't work properly after the snapshot is
restored as a new "VM".

Instead we do the hacky thing of just copy the update_settings.json file from the
old container to the new one. Docker doesn't seem to give us many options here.