Fix race condition in file upload by waiting for mkdir completion
Resolve an issue in the Upload method of fileService. Upload attempts
to ensure the directory exists for a file to be uploaded, but when the
"mkdir" was run via docker exec, the command was started asynchronously.

If the system were slow, this could mean CopyToContainer could run
before the directory was actually created leading to intermittent
failures during disk attach operations.